Convert all test files to TS and add .spec to all filenames
Change-type: patch
