+++ test/26-supervisor-api-auth.spec.ts
